rebuilding the model

+++ repository-jdbc/src/main/java/com/okarin/repository/DeveloperRepository.java
@@ -3,7 +3,6 @@ package com.okarin.repository;
 import com.okarin.entity.Developer;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
@@ -15,6 +14,6 @@ public interface DeveloperRepository extends CrudRepository<Developer,Long> {
     List<Developer> findAll();
 
     @Query(value = "SELECT * FROM developers where projects_id=?1 order by id",nativeQuery = true)
-    List<Developer> findByProjectId(Integer projectId);
+    List<Developer> findByProjectId(Long projectId);
 
 }
